Feature fridge content (#4)
* feat: implement fridge content layout

* feat: implement fridge product deleting

* feat: add fridgeId field in fridge products filter

* feat(card): make cards shrink on hover

* fix: fix Int overflow v2

* feat: implement new fridge content screen + viewModel

* feat: navigate to fridge on click

* feat(auth): finally fix auth interceptors with a separate refresh auth service

* chore: move some declarations

* feat: implement more realistic user agent

* fix: get fridge name for the content header

* fix: validate if fridge name is blank on fridge create

+++ app/src/main/java/com/haykor/fridge/core/data/local/datastore/TokenManager.kt
@@ -26,9 +26,9 @@ class TokenManager @Inject constructor(
 
     suspend fun saveTokens(authResponse: AuthResponse) {
         val accessTokenExpiryTime =
-            System.currentTimeMillis() + (authResponse.accessTokenExpiresIn * 60 * 1000L)
+            System.currentTimeMillis() + (authResponse.accessTokenExpiresIn * 60L * 1000L)
         val refreshTokenExpiryTime =
-            System.currentTimeMillis() + (authResponse.refreshTokenExpiresIn * 24 * 60 * 60 * 1000L)
+            System.currentTimeMillis() + (authResponse.refreshTokenExpiresIn.toLong() * 24L * 60L * 60L * 1000L)
 
         dataStore.edit { preferences ->
             preferences[ACCESS_TOKEN] = authResponse.accessToken
@@ -54,7 +54,7 @@ class TokenManager @Inject constructor(
 
     suspend fun isRefreshTokenExpired(): Boolean {
         val expiry =
-            dataStore.data.map { it[REFRESH_TOKEN_EXPIRY] ?: 0L }.first() // FIXME: doesn't work idk
+            dataStore.data.map { it[REFRESH_TOKEN_EXPIRY] ?: 0L }.first()
         Log.d("tokens", "expiryRefresh=$expiry currentTime=${System.currentTimeMillis()}")
         return System.currentTimeMillis() >= expiry
     }
